pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cfb-intel"
version = "0.1.0"
description = "College football player intelligence scraper and data pipeline"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=1.10,<2.0",
  "requests>=2.31",
  "beautifulsoup4>=4.12",
  "feedparser>=6.0",
  "python-dotenv>=1.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "scripts", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
